views.py -- Added methods experiment_list && experiment_detail
We added two methods called experiment_list && experiment_detail basically the reason we want to add these methods is the first method _list we want to be able to list all the objects in the **query-set** coming from out data that we implemented into our `models.py`, we want to import JsonResponse to because we want out data to return in JSon form when we to visit the website page `//localhost:7000/experiments` it will render the data in JSON form.  We also want the ability to update the data by using the POST method, then breaking down the data using JSONParser, then creating new data in the serializer as long as the data is valid it will save it with a http status of 201, if not then http status of 400.

Next method is the ability to Retrieve the data, update it or delete it.  using the try and except method we can try to get data and return a validation error is the data does not exist.(http status 404)