Update serializers.py
Creating a serializers.py to be able to convert models into JSON, serializers is apart of the rest_framwork and it works very similiar to Djangos Forms.py.  we created a serializer class to control the output of our responses as well as the ModelSerializer class that provides a useful shortcut creating serializers that deal with model instance and querysets.