Update models.py
We start by importing the Pygments get_all_lexers && get_all_styles we import these to split the keyword, string, or comment.  We have a LEXER for every language and markup.  we want to define LEXER as get_all_lexers which will return an iterable over all registered lexers in Pygments in a tuple format.  we want to display all languages as well but we want to sort them list of two.  we also have Styles_choices to display all the styles that are in registered in Pygments.  We have one model class to communicate to our DB to display created, title, code, linenos, language, and style.  we made created to be set to editable false then we define a save method to create a save in the DB for the field created.