Update views.py
by using mixin and generic classes we can use pre-built views that will allow us to write smaller code and get what we need done faster.  we would only need to call the objects once in the class and then we would use the *args and **kwargs to call the objects. and their values.