Update urls.py
As you can see our URLs have decreased in size by adding the router.  it allows you to call the ViewSet class we created to group all three separate Views and call that instead of calling each view and assigning a url to it.  DefaultRouter allows us to call the api_root basically displaying our root begin of our API, last we just need to include the router.urls into our urlpatterns for it to generate a url for each entry.  Then of course we need to call our Logins for the APIs.