Create __init__.py
empty file so we know this is a python base package.  we are going to create a project structure called collections to test the importing of a 3rd party API.  This is solely a test and this is still in the works of implementing it into the real site.  lets begin by adding some tests.