Added another model cls
we want to implement a test to communicate to the musicbrainzNGS API for pulling artist's and their releases.  we first want to define a model with certain fields we are going to use a @classmethod to define a method to pull artist and return the list of any artist we are looking for with their releases.  we have another method to pull certain tag-list, so we will only display artists that are map to the particular map name and counts we listed in that method.