Update tasks.py
here we are using the @shared_task to call the the get artist tracks from musicbrainz api.  we orignally had it in the models but by moving it tasks.py we can make this independent from our entire project. A simple and fast task queue for executing multiple tasks in Django without relying on your project.